// src/app/business/cirugias/interfaces/Tabla_cirugias.ts
export default interface Tabla_cirugias{
    id:number;
    fechaCirugia: String;
    numCirugia: Number;
    numExpediente: Number;
    nomPaciente: String;
    apellidoPPaciente: String;
    apellidoMPaciente: String;
    nomCirujano: String;
    nomAnestesiologo: String;
    nomResidente: String;
    nomCirugia: String;
    nomEnfeInt: String;
    nomEnfeCir: String;
    tipoAnes: String;
    iniciaAnes: string;
    finAnes: string;
    medicamento: string;
    suturas: string;
    material: string;
    validCheck: string;
}